Move safariFileSave to its own module for ease of testing
* Add test case for download when user is on safari
